assistant/pyproject.toml
[project]
name = "voice-assistant"
version = "0.1.0"
description = "Voice assistant pipeline orchestrating STT → LLM → TTS"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.2",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
